pyproject.toml
[build-system]
requires      = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "stegcore"
version         = "2.0.11"
description     = "Crypto-steganography toolkit with deniable payloads"
readme          = "README.md"
license         = "AGPL-3.0"
requires-python = ">=3.11"

dependencies = [
    "customtkinter>=5.2",
    "Pillow>=10.0",
    "numpy>=1.26",
    "ascon>=0.0.9",
    "cryptography>=41.0",
    "argon2-cffi>=23.1",
    "pyzstd>=0.15",
    "typer>=0.12",
    "rich>=13.0",
]

[project.scripts]
# CLI: stegcore embed / extract / score / info / ciphers / wizard
stegcore = "main:main"

[project.gui-scripts]
# GUI: no console window (Windows: no cmd popup; Linux/macOS: detaches terminal)
stegcore-gui = "main:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "Pillow>=10.0",
    "numpy>=1.26",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "--cov=core --cov-report=html:htmlcov --cov-report=term-missing --cov-fail-under=90"

[tool.coverage.run]
omit = ["core/utils.py"]

[tool.setuptools]
# top-level .py files that are not inside a package directory
py-modules = ["main", "cli"]

[tool.setuptools.packages.find]
where    = ["."]
exclude  = ["stegcore_venv*", "build*", "dist*", "tests*"]

[tool.setuptools.package-data]
"*" = ["*.ico", "*.png"]
